move mermaid to seperate file
